Extract testing boilerplate from distributed_queries.py e2e to common.py
