+++ tests/e2e/distributed_queries/CMakeLists.txt
@@ -3,3 +3,4 @@ function(distributed_queries_e2e_python_files FILE_NAME)
 endfunction()
 
 distributed_queries_e2e_python_files(distributed_queries.py)
+distributed_queries_e2e_python_files(common.py)
